chore: extract normalisers to file
